Update trait-geotarget-keys.md
User-defined keys section was unclear, and was misleading customers to believe "d_zx" was a pre-defined dimension for zip code. Further down the page, the actual pre-defined dimension "d_postal_code" is outlined. I just wanted to clarify the documentation that "d_zx" is just an example of a possible user-defined key in this case. Please feel free to re-word or alter my changes in anyway to grant more clarity.