Update cdf-files.md
Proposing changes to bring these documents up to date with the changes to CDF exports. I added in the new line file separator (/N), as we're observing this in customer exports. I'm not aware of the situations which this separator is utilized, or if the old separators are still in use, so I didn't make any changes there. I also added in the updates to the file names, which now include a file sequence for the various file fragments. I was able to confirm these details from the following email communication, forwarded to me by Shane Nielson:

What's Changing?
In summary, Audience Manager is migrating the backend processing for CDF jobs to an updated infrastructure to enhance overall stability. The migration has already been tested with a subset of customers with no meaningful impact to their operations to consume and process the files. 
 
What Do Customers Need to Do?
Customers are not required to make any changes to their existing operations. An overview of potential changes in the CDF jobs that may be observed are outlined below:
 
1.	Request_parameters:
The request_params field gets encoded using java.net.URLEncoder currently but will transition to JavaScript (rather than Java). JS's encodeURIComponent is used. For example, the space character is encoded as '%20' currently but will be encoded as '+.' 
 
2.	Change in number of files being written to S3
The total number of rows will be the same but the way the data is split into files will be different and is a function of the warehouse size
 
3.	Filename format: 
AAM_CDF_<PID>_xxxxxx_x.gz will now look like AAM_CDF_<PID>_y_y_y.gz. For example, a file generated by current methods would be named as AAM_CDF_8058_000000_0.gz while a new generated file will be named as AAM_CDF_8058_7_7_0.gz.